Change log entries for user tags on volumes and updating subnets on ALBs (GA, 2022-08-23)
The `publish` branch PR currently includes the change log entry for user tags for volumes (from merged #1457) only. Depending on team feedback in https://github.ibm.com/riaas/api-spec/pull/4108, the plan is to also include the change log entry for updating subnets on ALBs (from merged PR #1441 and in-progress update to that entry via #1482). Note that 1482 is blocked by 4108.